problem solution explanation
We can solve this problem using C programming. For Linux operating systems we use pthread library to create,manage and destroy threads and mutex locks. Mutex lock is a function that ensures that only one out of all the members who are trying to access shared variable will be accessing it. Threads are in general a process in a program.